requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.35.2
git+https://github.com/openai/CLIP.git
Pillow==10.1.0
opencv-python-headless==4.8.1.78
moviepy==1.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
accelerate==0.24.1